% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predict.R
\name{anova.modeler}
\alias{anova.modeler}
\title{Extra Sum-of-Squares F-Test for objects of class \code{modeler}}
\usage{
\method{anova}{modeler}(reduced_model, full_model = NULL, ...)
}
\arguments{
\item{reduced_model}{An object of class `modeler` with reduced number of parameters.}

\item{full_model}{An object of class `modeler` with more number of parameters.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A tibble with columns giving F test and p values.
}
\description{
anova for objects of class \code{modeler}
}
\examples{
library(exploreHTP)
dt <- data.frame(X = 1:6, Y = c(12, 16, 44, 50, 95, 100))
f1 <- function(x, b, m) {
  y <- b + m * x
  return(y)
}
f2 <- function(x, a, b, c) {
  y <- a + b * x + c * x^2
  return(y)
}
mo_1 <- modeler(dt, x = X, y = Y, fn = "f1", param = c(b = -5, m = 10))
mo_2 <- modeler(dt, x = X, y = Y, fn = "f2", param = c(a = 1, b = 10, c = 5))
anova(mo_1, mo_2)
}
\author{
Johan Aparicio [aut]
}
