% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predict.R
\name{coef.modeler}
\alias{coef.modeler}
\title{Coefficients of an object of class \code{modeler}}
\usage{
\method{coef}{modeler}(x, id = NULL, metadata = FALSE, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{id}{A numeric Id to filter by.}

\item{metadata}{TRUE or FALSE. Whether to bring the metadata or not when calculating the coefficients.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A data.frame object with coefficients and standard errors.
}
\description{
Coefficients for an object of class \code{modeler}
}
\examples{
library(exploreHTP)
data(dt_potato)
explorer <- read_HTP(dt_potato, x = DAP, y = c(Canopy, GLI_2), id = Plot)
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    by = Plot,
    id = c(15, 2, 45),
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    fn = "fn_piwise",
    add_zero = TRUE,
    max_as_last = TRUE
  )
mod_1
coef(mod_1, id = 2)
}
\author{
Johan Aparicio [aut]
}
