% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_predict.R
\name{coef.modeler_HTP}
\alias{coef.modeler_HTP}
\title{Coefficients of an object of class \code{modeler_HTP}}
\usage{
\method{coef}{modeler_HTP}(x, id = NULL, metadata = FALSE, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler_HTP} resulting of
executing the function \code{modeler_HTP()}}

\item{id}{A numeric to filter by Plot Id.}

\item{metadata}{TRUE or FALSE. Whether to bring the metadata or not when calculating the coefficients.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A data.frame object with coefficients and standard errors.
}
\description{
Coefficients for an object of class \code{modeler_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- dt_potato |>
  read_HTP(
    x = DAP,
    y = c(Canopy, GLI_2),
    id = Plot,
    .keep = c(Gen, Row, Range)
  )
mod <- modeler_HTP(
  x = results,
  index = "Canopy",
  id = 1:2,
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise"
)
mod
coef(mod)
}
\author{
Johan Aparicio [aut]
}
