% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predict.R
\name{confint.modeler}
\alias{confint.modeler}
\title{Confidence Intervals for an object of class \code{modeler}}
\usage{
\method{confint}{modeler}(x, parm = NULL, level = 0.95, id = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{parm}{A specification of which parameters are to be given confidence intervals, must be a vector of names. If missing, all parameters are considered.}

\item{level}{The confidence level required. Default is 0.95.}

\item{id}{A unique identifier to filter by. \code{NULL} by default.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A tibble with columns giving lower and upper confidence limits for each parameter.
}
\description{
confint for an object of class \code{modeler}
}
\examples{
library(exploreHTP)
data(dt_potato)
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    subset = c(15, 35, 45),
    add_zero = TRUE,
    max_as_last = TRUE
  )
print(mod_1)
confint(mod_1)
}
\author{
Johan Aparicio [aut]
}
