% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dt_potato.R
\docType{data}
\name{dt_chips}
\alias{dt_chips}
\title{Drone Data in Potato 2022}
\format{
A tibble with 1568 rows and 17 variables:
\describe{
  \item{Trial}{chr trial name}
  \item{DAP}{dbl denoting Days after planting}
  \item{Plot}{dbl denoting the unique plot id}
  \item{Row}{dbl denoting the row coordinate}
  \item{Range}{dbl denoting range coordinate}
  \item{Gen}{chr denoting the genotype id}
  \item{Yield}{dbl Total Yield measured at the end of the season}
  \item{VineMaturity}{dbl VineMaturity measure once during the season}
  \item{Red - Canopy}{dbl denoting the drone-derived traits}
}
}
\source{
UW - Potato Breeding Program
}
\usage{
dt_chips
}
\description{
Vegetative indices and biomass related traits for a potato trial arranged in a p-rep design.
}
\keyword{datasets}
