% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_dt_potato.R
\docType{data}
\name{dt_potato_22}
\alias{dt_potato_22}
\title{Drone-derived data from a potato breeding trial (2022)}
\format{
A tibble with 1764 rows and 9 variables:
\describe{
  \item{Trial}{chr trial name}
  \item{Plot}{dbl denoting the unique plot id}
  \item{Row}{dbl denoting the row coordinate}
  \item{Range}{dbl denoting range coordinate}
  \item{gid}{chr denoting the genotype id}
  \item{DAP}{dbl denoting Days after planting}
  \item{GC}{dbl UAV-Derived Ground Cover}
  \item{PH}{dbl UAV-Derived Plant Height}
  \item{Yield}{Total Yield}
}
}
\source{
UW - Potato Breeding Program 2022
}
\usage{
dt_potato_22
}
\description{
Ground cover and plant height for a potato trial arranged in a p-rep design.
}
\keyword{datasets}
