% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexFitR-package.R
\docType{package}
\name{flexFitR-package}
\alias{flexFitR}
\alias{flexFitR-package}
\title{flexFitR: Flexible Non-Linear Least Square Model Fitting in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements general-purpose optimization techniques for non-linear least squares problems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://apariciojohan.github.io/flexFitR/}
  \item \url{https://github.com/AparicioJohan/flexFitR}
  \item Report bugs at \url{https://github.com/AparicioJohan/flexFitR/issues}
}

}
\author{
\strong{Maintainer}: Johan Aparicio \email{aparicioarce@wisc.edu}

Authors:
\itemize{
  \item Jeffrey Endelman \email{endelman@wisc.edu}
}

Other contributors:
\itemize{
  \item University of Wisconsin Madison [copyright holder]
}

}
\keyword{internal}
