% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_canopy.R
\name{fn_canopy}
\alias{fn_canopy}
\title{Piece-wise Regression}
\usage{
fn_canopy(t, t1 = 45, t2 = 80, k = 0.9)
}
\arguments{
\item{t}{Numeric value.}

\item{t1}{First break point.}

\item{t2}{t value to reach the maximum value.}

\item{k}{Maximum y value.}
}
\value{
Numeric value.
}
\description{
Piece-wise Regression
}
\examples{
library(exploreHTP)
x <- seq(0, 108, 0.1)
y_hat <- sapply(x, FUN = fn_canopy, t1 = 34.9, t2 = 61.8, k = 100)
plot(x, y_hat, type = "l")
lines(x, y_hat, col = "red")
abline(v = c(34.9, 61.8), lty = 2)
}
