% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_growth_curves.R
\name{fn_lin_pl_lin2}
\alias{fn_lin_pl_lin2}
\title{Linear Plateau Linear with Constrains}
\usage{
fn_lin_pl_lin2(t, t1, dt, t3, k, beta)
}
\arguments{
\item{t}{Numeric. The time value.}

\item{t1}{Numeric. The lower threshold time. Default is 45.}

\item{dt}{Numeric. dt = t3 - t2. Default is 28}

\item{t3}{Numeric. The lower threshold time after plateau. Default is 45.}

\item{k}{Numeric. The maximum value of the function. Default is 0.9.}

\item{beta}{Numeric. Slope of the linear decay.}
}
\value{
A numeric value based on the linear plateau linear model.
}
\description{
Linear Plateau Linear with Constrains
}
\details{
\if{html}{
\deqn{
f(t; t_1, dt, t_3, k, \beta) =
\begin{cases}
0 & \text{if } t < t_1 \\
\dfrac{k}{(t_3 - dt) - t_1} \cdot (t - t_1) & \text{if } t_1 \leq t \leq (t_3 - dt) \\
k & \text{if } (t_3 - dt) \leq t \leq t_3 \\
k + \beta \cdot (t - t_3) & \text{if } t > t_3
\end{cases}
}
}
}
\examples{
library(exploreHTP)
t <- seq(0, 108, 0.1)
y_hat <- sapply(
  X = t,
  FUN = fn_lin_pl_lin2,
  t1 = 38.7, dt = 28, t3 = 90, k = 0.32, beta = -0.01
)
plot(t, y_hat, type = "l")
lines(t, y_hat, col = "red")
abline(v = c(38.7, 62), lty = 2)
}
