% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_growth_curves.R
\name{fn_lin_pl_lin4}
\alias{fn_lin_pl_lin4}
\title{Linear Plateau Linear Constrains}
\usage{
fn_lin_pl_lin4(t, t1, t2, t3, k, beta)
}
\arguments{
\item{t}{Numeric. The time value.}

\item{t1}{Numeric. The lower threshold time. Default is 45.}

\item{t2}{Numeric. The upper threshold time before plateau. Default is 80.}

\item{t3}{Numeric. The lower threshold time after plateau. Default is 45.}

\item{k}{Numeric. The maximum value of the function. Default is 0.9.}

\item{beta}{Numeric. Slope of the linear decay.}
}
\value{
A numeric value based on the linear plateau linear model.
}
\description{
Linear Plateau Linear Constrains
}
\details{
\if{html}{
\deqn{
f(t; t_1, t_2, t_3, k, \beta) =
\begin{cases}
0 & \text{if } t < t_1 \\
\dfrac{k}{t_2 - t_1} \cdot (t - t_1) & \text{if } t_1 \leq t \leq t_2 \\
k & \text{if } t_2 \leq t \leq t_3 \\
k + \beta \cdot (t - t_3) & \text{if } t > t_3
\end{cases}
}
}
}
\examples{
library(exploreHTP)
t <- seq(0, 108, 0.1)
y_hat <- sapply(
  X = t,
  FUN = fn_lin_pl_lin,
  t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01
)
plot(t, y_hat, type = "l")
lines(t, y_hat, col = "red")
abline(v = c(38.7, 62), lty = 2)
}
