% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_growth_curves.R
\name{fn_logistic}
\alias{fn_logistic}
\title{Logistic function}
\usage{
fn_logistic(t, a, t0, k)
}
\arguments{
\item{t}{Numeric value.}

\item{a}{Numeric value.}

\item{t0}{Numeric value.}

\item{k}{Numeric value.}
}
\value{
A numeric value based on the logistic function.
}
\description{
Computes a value based on a logistic function.
}
\details{
\if{html}{
\deqn{
f(t; a, t0, k) = \frac{k}{1 + e^{-a(t - t_0)}}
}
}
}
\examples{
library(flexFitR)
plot_fn(
  fn = "fn_logistic",
  params = c(a = 0.199, t0 = 47.7, k = 100),
  interval = c(0, 108),
  n_points = 2000
)
}
