% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_growth_curves.R
\name{fn_piwise}
\alias{fn_piwise}
\title{Piecewise Regression}
\usage{
fn_piwise(t, t1 = 45, t2 = 80, k = 0.9)
}
\arguments{
\item{t}{Numeric. The time value.}

\item{t1}{Numeric. The lower threshold time. Default is 45.}

\item{t2}{Numeric. The upper threshold time. Default is 80.}

\item{k}{Numeric. The maximum value of the function. Default is 0.9. Assumed to be known.}
}
\value{
A numeric value based on the threshold model.
If \code{t} is less than \code{t1}, the function returns 0.
If \code{t} is between \code{t1} and \code{t2} (inclusive),
the function returns a value between 0 and \code{k} in a linear trend.
If \code{t} is greater than \code{t2}, the function returns \code{k}.
}
\description{
Piecewise Regression
}
\details{
\if{html}{
\deqn{
f(t; t_1, t_2, k) =
\begin{cases}
0 & \text{if } t < t_1 \\
\dfrac{k}{t_2 - t_1} \cdot (t - t_1) & \text{if } t_1 \leq t \leq t_2 \\
k & \text{if } t > t_2
\end{cases}
}
}
}
\examples{
library(exploreHTP)
t <- seq(0, 108, 0.1)
y_hat <- sapply(t, FUN = fn_piwise, t1 = 34.9, t2 = 61.8, k = 100)
plot(t, y_hat, type = "l")
lines(t, y_hat, col = "red")
abline(v = c(34.9, 61.8), lty = 2)
}
