% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_inverse_predict.R
\name{inverse_predict}
\alias{inverse_predict}
\title{Generic for inverse prediction}
\usage{
inverse_predict(object, ...)
}
\arguments{
\item{object}{An object for which to compute the inverse prediction.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic for inverse prediction
}
\keyword{internal}
