% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_growth_curves.R
\name{list_funs}
\alias{list_funs}
\title{Print available functions in flexFitR}
\usage{
list_funs()
}
\value{
A vector with available functions
}
\description{
Print available functions in flexFitR
}
\examples{
library(flexFitR)
list_funs()
}
