% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_metrics.R
\name{metrics_HTP}
\alias{metrics_HTP}
\title{Metrics for modeler_HTP}
\usage{
metrics_HTP(x, .by_plot = TRUE)
}
\arguments{
\item{x}{An object of class `modeler_HTP` containing the necessary data to compute the metrics.}

\item{.by_plot}{Return the metrics by plot? TRUE by default.}
}
\value{
A data frame containing the calculated metrics grouped by plot, row, range, genotype, and trait.
}
\description{
Computes various performance metrics for a modeler_HTP object.
The function calculates Sum of Squared Errors (SSE), Mean Absolute Error (MAE),
Mean Squared Error (MSE), Root Mean Squared Error (RMSE), and the Coefficient
of Determination (R-squared).
}
\details{
\if{html}{
Sum of Squared Errors (SSE):
\deqn{SSE = \sum_{i=1}^{n} (y_i - \hat{y}_i)^2}

Mean Absolute Error (MAE):
\deqn{MAE = \frac{1}{n} \sum_{i=1}^{n} |y_i - \hat{y}_i|}

Mean Squared Error (MSE):
\deqn{MSE = \frac{1}{n} \sum_{i=1}^{n} (y_i - \hat{y}_i)^2}

Root Mean Squared Error (RMSE):
\deqn{RMSE = \sqrt{\frac{1}{n} \sum_{i=1}^{n} (y_i - \hat{y}_i)^2}}

Coefficient of Determination (R-squared):
\deqn{R^2 = 1 - \frac{\sum_{i=1}^{n} (y_i - \hat{y}_i)^2}{\sum_{i=1}^{n} (y_i - \bar{y})^2}}
}
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- dt_potato |>
  read_HTP(
    x = DAP,
    y = c(Canopy, PH),
    id = Plot,
    .keep = c(Gen, Row, Range)
  )
x <- canopy_HTP(x = results, index = "Canopy", id = c(1:2))
plot(x, c(1:2))
print(x)
metrics_HTP(x)
}
