% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_modeler.R
\name{modeler_HTP}
\alias{modeler_HTP}
\title{Modeler HTP}
\usage{
modeler_HTP(
  results,
  index = "GLI",
  plot_id = NULL,
  check_negative = TRUE,
  add_zero = TRUE,
  method = c("subplex", "pracmanm", "anms"),
  return_method = FALSE,
  parameters = NULL,
  lower = -Inf,
  upper = Inf,
  initial_vals = NULL,
  fn = "fn_piwise",
  n_points = 1000,
  max_time = NULL
)
}
\arguments{
\item{results}{An object of class \code{read_HTP}, containing the results of the \code{read_HTP()} function.}

\item{index}{A string specifying the trait to be modeled. Default is \code{"GLI"}.}

\item{plot_id}{An optional vector of plot IDs to filter the data. Default is \code{NULL}, meaning all plots are used.}

\item{check_negative}{Logical. Convert negative values to zero. \code{TRUE} by default.}

\item{add_zero}{\code{TRUE} or \code{FALSE}. Add zero to the time series.\code{TRUE} by default.}

\item{method}{A vector of the methods to be used, each as a character string.
See optimx package. c("subplex", "pracmanm", "anms") by default.}

\item{return_method}{\code{TRUE} or \code{FALSE}. To return the method selected for the
optimization in the table. \code{FALSE} by default.}

\item{parameters}{Initial values for the parameters to be optimized over. c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01) by default.}

\item{lower}{Bounds on the variables for methods such as "L-BFGS-B" that can handle box (or bounds) constraints.}

\item{upper}{Bounds on the variables for methods such as "L-BFGS-B" that can handle box (or bounds) constraints.}

\item{initial_vals}{data.frame with columns <plot, genotype, t1, t2, ... , and all the initial parameters>.}

\item{fn}{String character with the name of the function "fn_lin_pl_lin".}

\item{n_points}{Number of time points to approximate the AUC. 1000 by default.}

\item{max_time}{Maximum time value for calculating the AUC. \code{NULL} by default takes the last time point.}
}
\value{
An object of class \code{modeler_HTP}, which is a list containing the following elements:
\describe{
  \item{\code{param}}{A data frame containing the optimized parameters and related information.}
  \item{\code{dt}}{A data frame with data used.}
  \item{\code{fn}}{The call used to calculate the AUC.}
  \item{\code{max_time}}{Maximum time value used for calculating the AUC.}
}
}
\description{
Modeler HTP
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "GLI_2"),
  row = "Row",
  range = "Range"
)
names(results)
mat <- modeler_HTP(
  results = results,
  index = "GLI_2",
  plot_id = c(195, 40),
  parameters = c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01),
  fn = "fn_lin_pl_lin",
)
plot(mat, plot_id = c(195, 40))
mat$param
can <- modeler_HTP(
  results = results,
  index = "Canopy",
  plot_id = c(195, 40),
  parameters = c(t1 = 45,t2 = 80, k = 0.9),
  fn = "fn_piwise",
)
plot(can, plot_id = c(195, 40))
can$param
}
