% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.canopy_HTP}
\alias{plot.canopy_HTP}
\title{Plot an object of class \code{canopy_HTP}}
\usage{
\method{plot}{canopy_HTP}(x, plot_id = NULL, label_size = 4, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{canopy_HTP} resulting of
executing the function \code{canopy_HTP()}}

\item{plot_id}{To avoid to many plots in one figure. Filter by Plot Id.}

\item{label_size}{Label size. 3 by default.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{canopy_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
print(results)
out <- canopy_HTP(results, plot_id = 22)
plot(out)
}
\author{
Johan Aparicio [aut]
}
