% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.height_HTP}
\alias{plot.height_HTP}
\title{Plot an object of class \code{height_HTP}}
\usage{
\method{plot}{height_HTP}(x, plot_id = NULL, label_size = 4, base_size = 14, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{height_HTP} resulting of
executing the function \code{height_HTP()}}

\item{plot_id}{To avoid too many plots in one figure. Filter by Plot Id.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{height_HTP}
}
\examples{
library(exploreHTP)
data(dt_chips)
results <- read_HTP(
  data = dt_chips,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(60, 150),
  correct_max = TRUE,
  add_zero = TRUE
)
names(out)
plot(out, plot_id = c(60, 150))
ph_1 <- height_HTP(
  results = results,
  canopy = out,
  plant_height = "PH",
  add_zero = TRUE,
  method = c("nlminb", "anms", "mla", "pracmanm", "subplex"),
  return_method = TRUE,
  parameters = c(t2 = 67, alpha = 1 / 600, beta = -1 / 80),
  fn_sse = sse_exp2_exp,
  fn = quote(fn_exp2_exp(time, t1, t2, alpha, beta))
)
plot(x = ph_1, plot_id = c(60, 150))
ph_1$param

ph_2 <- height_HTP(
  results = results,
  canopy = out,
  plant_height = "PH",
  add_zero = TRUE,
  method = c("nlminb", "anms", "mla", "pracmanm", "subplex"),
  return_method = TRUE,
  parameters = c(t2 = 67, alpha = 1 / 600, beta = -1 / 80),
  fn_sse = sse_exp2_lin,
  fn = quote(fn_exp2_lin(time, t1, t2, alpha, beta))
)
plot(x = ph_2, plot_id = c(60, 150))
ph_2$param
}
\author{
Johan Aparicio [aut]
}
