% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.maturity_HTP}
\alias{plot.maturity_HTP}
\title{Plot an object of class \code{maturity_HTP}}
\usage{
\method{plot}{maturity_HTP}(x, plot_id = NULL, label_size = 4, base_size = 14, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{maturity_HTP} resulting of
executing the function \code{maturity_HTP()}}

\item{plot_id}{To avoid too many plots in one figure. Filter by Plot Id.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{maturity_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "GLI_2"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(195, 40),
  correct_max = TRUE,
  add_zero = TRUE
)
mat <- maturity_HTP(
  results = results,
  canopy = out,
  index = "GLI_2",
  parameters = c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01),
  fn_sse = sse_lin_pl_lin,
  fn = quote(fn_lin_pl_lin(time, t1, t2, t3, k, beta))
)
plot(mat, plot_id = c(195, 40))
mat$param
}
\author{
Johan Aparicio [aut]
}
