% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_plots.R
\name{plot.modeler}
\alias{plot.modeler}
\title{Plot an object of class \code{modeler}}
\usage{
\method{plot}{modeler}(x, id = NULL, label_size = 4, base_size = 14, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{id}{To avoid too many plots in one figure. Filter by group Id.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{modeler}
}
\examples{
library(exploreHTP)
suppressMessages(library(dplyr))
data(dt_potato)
explorer <- explorer(dt_potato, x = DAP, y = c(Canopy, GLI_2), id = Plot)
# Example 1
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = GLI_2,
    grp = Plot,
    id = c(1:2),
    fn = "fn_lin_pl_lin",
    parameters = c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01),
    add_zero = TRUE,
    max_as_last = TRUE
  )
plot(mod_1, plot_id = c(195))
mod_1
# Example 2
mod_2 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    id = c(1:2),
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    add_zero = TRUE,
    max_as_last = TRUE
  )
plot(mod_2, plot_id = c(195))
mod_2
# Example 3
fixed_params <- explorer$dt_long |>
  filter(var \%in\% "Canopy") |>
  group_by(uid) |>
  summarise(k = max(y, na.rm = TRUE), .groups = "drop")
mod_3 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    id = 195,
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    fixed_params = fixed_params,
    add_zero = TRUE,
    max_as_last = TRUE
  )
plot(mod_3, id = c(195))
mod_3
}
\author{
Johan Aparicio [aut]
}
