% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.read_HTP}
\alias{plot.read_HTP}
\title{Plot an Object of Class \code{read_HTP}}
\usage{
\method{plot}{read_HTP}(
  x,
  type = "trait_by_time",
  label_size = 4,
  signif = FALSE,
  method = "pearson",
  filter_trait = NULL,
  n_row = NULL,
  n_col = NULL,
  base_size = 13,
  return_gg = FALSE,
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{read_HTP}, resulting from executing the function \code{read_HTP()}.}

\item{type}{Character string specifying the type of plot to generate. Available options are:
\describe{
  \item{\code{"trait_by_time"}}{Plots correlations between traits over time (default).}
  \item{\code{"time_by_trait"}}{Plots correlations between time points for each trait.}
  \item{\code{"evolution"}}{Plots the evolution of traits over time.}
}}

\item{label_size}{Numeric. Size of the labels in the plot. Default is 4.}

\item{signif}{Logical. If \code{TRUE}, adds p-values to the correlation plot labels. Default is \code{FALSE}.}

\item{method}{Character string specifying the method for correlation calculation. Available options are \code{"pearson"} (default), \code{"spearman"}, and \code{"kendall"}.}

\item{filter_trait}{Character vector specifying the traits to exclude from the plot.}

\item{n_row}{Integer specifying the number of rows to use in \code{facet_wrap()}. Default is \code{NULL}.}

\item{n_col}{Integer specifying the number of columns to use in \code{facet_wrap()}. Default is \code{NULL}.}

\item{base_size}{Numeric. Base font size for the plot. Default is 13.}

\item{return_gg}{Logical. If \code{TRUE}, returns the ggplot object instead of printing it. Default is \code{FALSE}.}

\item{...}{Further graphical parameters for future improvements.}
}
\value{
A ggplot object and an invisible data.frame containing the correlation table when \code{type} is \code{"trait_by_time"} or \code{"time_by_trait"}.
}
\description{
Creates various plots for an object of class \code{read_HTP}. Depending on the specified type, the function can generate plots that show correlations between traits over time, correlations between time points for each trait, or the evolution of traits over time.
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
table <- plot(results, label_size = 4, signif = TRUE, n_row = 2)
table
plot(results, type = "time_by_trait", label_size = 4, signif = TRUE)
}
