% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_predict.R
\name{predict.modeler_HTP}
\alias{predict.modeler_HTP}
\title{Predict an object of class \code{modeler_HTP}}
\usage{
\method{predict}{modeler_HTP}(x, time = NULL, plot_id = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler_HTP} resulting of
executing the function \code{modeler_HTP()}}

\item{time}{A numeric time point to make the prediction. Can be more than one.}

\item{plot_id}{A numeric to filter by Plot Id.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A data.frame object with predicted values.
}
\description{
Model predictions for an object of class \code{modeler_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "GLI_2"),
  row = "Row",
  range = "Range"
)
mod <- modeler_HTP(
  x = results,
  index = "Canopy",
  plot_id = c(15, 2, 45),
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise"
)
mod
predict(mod, time = 45, plot_id = 2)
}
\author{
Johan Aparicio [aut]
}
