% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_print.R
\name{print.modeler_HTP}
\alias{print.modeler_HTP}
\title{Print an object of class \code{modeler_HTP}}
\usage{
\method{print}{modeler_HTP}(x, ...)
}
\arguments{
\item{x}{An object fitted with the function \code{modeler_HTP()}.}

\item{...}{Options used by the tibble package to format the output. See
`tibble::print()` for more details.}
}
\value{
an object inheriting from class \code{modeler_HTP}.
}
\description{
Prints information about \code{modeler_HTP} function.
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- dt_potato |>
  read_HTP(
    x = DAP,
    y = c(Canopy, PH),
    id = Plot,
    .keep = c(Gen, Row, Range)
  )
mod_1 <- modeler_HTP(
  x = results,
  variable = "Canopy",
  id = c(1:5),
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise",
  max_as_last = TRUE
)
plot(mod_1, id = c(1:5))
print(mod_1)
}
\author{
Johan Aparicio [aut]
}
