% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_canopy.R
\name{sse}
\alias{sse}
\title{Sum of Squares Error Function}
\usage{
sse(params, t, y)
}
\arguments{
\item{params}{Numeric vector with two parameters.}

\item{t}{Independent variable.}

\item{y}{Response variable.}
}
\value{
sum of squares error
}
\description{
Function to be minimized in the optimx package.
}
\examples{
library(exploreHTP)
x <- c(0, 29, 36, 42, 56, 76, 92, 100, 108)
y <- c(0, 0, 4.379, 26.138, 78.593, 100, 100, 100, 100)
sse(params = c(34.9, 61.8), t = x, y = y)

y_hat <- sapply(x, FUN = fn_canopy, t1 = 34.9, t2 = 61.8, k = 100)
sum((y - y_hat)^2)
}
