% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_growth_curves.R
\name{sse_exp1_lin}
\alias{sse_exp1_lin}
\title{Sum of Squares Error Function for Exponential Linear Model 1}
\usage{
sse_exp1_lin(params, t, y, t1)
}
\arguments{
\item{params}{Numeric vector. The parameters for the \link{fn_exp1_lin}
function, where \code{params[1]} is \code{t2}, \code{params[2]} is \code{alpha},
and \code{params[3]} is \code{beta}.}

\item{t}{Numeric vector. The time values.}

\item{y}{Numeric vector. The observed values.}

\item{t1}{Numeric. The lower threshold time.}
}
\value{
A numeric value representing the sum of squared errors.
}
\description{
Calculates the sum of squared errors (SSE) between observed values and values
predicted by the \link{fn_exp1_lin} function.
}
\examples{
params <- c(50, 0.1, -0.01)
t <- c(10, 20, 30, 40, 50, 60)
y <- c(0, 0, 0.2, 0.5, 0.8, 0.9)
t1 <- 20
sse_exp1_lin(params, t, y, t1) # Should return the SSE value

}
