% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_growth_curves.R
\name{sse_lin_pl_lin2}
\alias{sse_lin_pl_lin2}
\title{Sum of Squares Error Function for Linear Plateau Linear Model with Constrains}
\usage{
sse_lin_pl_lin2(params, t, y)
}
\arguments{
\item{params}{Numeric vector. The parameters for the \code{fn_lin_pl_lin2} function,
where \code{params[1]} is \code{t1}, \code{params[2]} is \code{t2}, \code{params[3]} is \code{t3}
\code{params[4]} is \code{k} and \code{params[5]} is \code{beta}.}

\item{t}{Numeric vector. The time values.}

\item{y}{Numeric vector. The observed values.}
}
\value{
A numeric value representing the sum of squared errors.
}
\description{
Calculates the sum of squared errors (SSE) between observed values and values
predicted by the \link{fn_lin_pl_lin2} function. This is the objective function to
be minimized in the optimx package.
}
\examples{
library(exploreHTP)
x <- c(0, 29, 36, 42, 56, 76, 92, 100, 108)
y <- c(0, 0, 0, 0.027, 0.185, 0.325, 0.321, 0.256, 0.176)
sse_lin_pl_lin2(
  params = c(t1 = 38.7, dt = 28, t3 = 90, k = 0.32, beta = -0.01),
  t = x,
  y = y
)
y_hat <- sapply(
  X = x,
  FUN = fn_lin_pl_lin2,
  t1 = 38.7, dt = 28, t3 = 90, k = 0.32, beta = -0.01
)
sum((y - y_hat)^2)
}
