% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_growth_curves.R
\name{sse_piwise}
\alias{sse_piwise}
\title{Sum of Squares Error Function for Piecewise Model}
\usage{
sse_piwise(params, t, y)
}
\arguments{
\item{params}{Numeric vector. The parameters for the \code{fn_piwise} function,
where \code{params[1]} is \code{t1} and \code{params[2]} is \code{t2}.}

\item{t}{Numeric vector. The time values.}

\item{y}{Numeric vector. The observed values.}
}
\value{
A numeric value representing the sum of squared errors.
}
\description{
Calculates the sum of squared errors (SSE) between observed values and values
predicted by the \link{fn_piwise} function. This is the objective function to
be minimized in the optimx package.
}
\examples{
library(exploreHTP)
x <- c(0, 29, 36, 42, 56, 76, 92, 100, 108)
y <- c(0, 0, 4.379, 26.138, 78.593, 100, 100, 100, 100)
sse_piwise(params = c(34.9, 61.8), t = x, y = y)

y_hat <- sapply(x, FUN = fn_piwise, t1 = 34.9, t2 = 61.8, k = 100)
sum((y - y_hat)^2)
}
