% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predict.R
\name{vcov.modeler}
\alias{vcov.modeler}
\title{Variance-Covariance matrix for an object of class \code{modeler}}
\usage{
\method{vcov}{modeler}(x, id = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{id}{A unique identifier to filter by. \code{NULL} by default.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A list object with matrices of the estimated covariances between the parameter estimates.
}
\description{
vcov for an object of class \code{modeler}
}
\examples{
library(flexFitR)
data(dt_potato)
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    subset = c(15, 2, 45),
    add_zero = TRUE,
    max_as_last = TRUE
  )
print(mod_1)
vcov(mod_1)
}
\author{
Johan Aparicio [aut]
}
