% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predict.R
\name{vcov.modeler_HTP}
\alias{vcov.modeler_HTP}
\title{Variance-Covariance matrix for an object of class \code{modeler_HTP}}
\usage{
\method{vcov}{modeler_HTP}(x, id = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler_HTP} resulting of
executing the function \code{modeler_HTP()}}

\item{id}{A numeric Id to filter by.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A list object with matrices of the estimated covariances between the parameter estimates.
}
\description{
vcov for an object of class \code{modeler_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
mod_1 <- dt_potato |>
  modeler_HTP(
    x = DAP,
    y = Canopy,
    by = Plot,
    id = c(15, 2, 45),
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    fn = "fn_piwise",
    max_as_last = TRUE
  )
mod_1
vcov(mod_1)
}
\author{
Johan Aparicio [aut]
}
