% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_canopy.R
\name{canopy_HTP}
\alias{canopy_HTP}
\title{Canopy Modelling}
\usage{
canopy_HTP(
  results,
  canopy = "Canopy",
  plot_id = NULL,
  correct_max = TRUE,
  add_zero = TRUE,
  method = c("subplex", "pracmanm", "anms"),
  return_method = FALSE,
  parameters = c(t1 = 45, t2 = 80),
  fn = fn_sse_can
)
}
\arguments{
\item{results}{An object of class \code{exploreHTP}, containing the results of the \code{read_HTP()} function.}

\item{canopy}{A string specifying the canopy trait to be modeled. Default is \code{"Canopy"}.}

\item{plot_id}{An optional vector of plot IDs to filter the data. Default is \code{NULL}, meaning all plots are used.}

\item{correct_max}{Logical. If \code{TRUE}, adds the maximum value after reaching the local maximum. Default is \code{TRUE}.}

\item{add_zero}{Logical. If \code{TRUE}, adds a zero value to the time series. Default is \code{TRUE}.}

\item{method}{A character vector of optimization methods to be used, as specified in the \code{optimx} package. Default is \code{c("subplex", "pracmanm", "anms")}.}

\item{return_method}{Logical. If \code{TRUE}, includes the selected optimization method in the output table. Default is \code{FALSE}.}

\item{parameters}{A named vector specifying the initial values for the parameters to be optimized. Default is \code{c(t1 = 45, t2 = 80)}.}

\item{fn}{A function to be minimized (or maximized), with the first argument being the vector of parameters over which minimization is to take place. It should return a scalar result. Default is \link{fn_sse_can}.}
}
\value{
A list with two elements:
\describe{
  \item{\code{param}}{A data frame containing the optimized parameters and related information.}
  \item{\code{dt}}{A data frame with the corrected and possibly zero-augmented canopy data.}
}
}
\description{
This function performs canopy modelling based on time series data from high-throughput phenotyping (HTP). It optimizes parameters to fit a specified function to the canopy data over time, potentially correcting maximum values and adding a zero point to the series.
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(22, 40),
  correct_max = TRUE,
  add_zero = TRUE
)
names(out)
plot(out, c(22, 40))
out$param$deltaT <- out$param$t2 - out$param$t1
out$param$slope <- out$param$max / out$param$deltaT
out$param
}
