% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_canopy.R
\name{canopy_HTP}
\alias{canopy_HTP}
\title{Canopy Modelling}
\usage{
canopy_HTP(
  results,
  canopy = "Canopy",
  plot_id = NULL,
  correct_max = TRUE,
  add_zero = TRUE,
  return_method = FALSE,
  method = c("subplex", "pracmanm", "anms")
)
}
\arguments{
\item{results}{Object of class exploreHTP}

\item{canopy}{string}

\item{plot_id}{Optional Plot ID. NULL by default}

\item{correct_max}{Add maximum value after reaching the local maximum. TRUE
by default.}

\item{add_zero}{TRUE or FALSE. Add zero to the time series.TRUE by default.}

\item{return_method}{TRUE or FALSE. To return the method selected for the
optimization in the table. FALSE by default.}

\item{method}{A vector of the methods to be used, each as a character string.
See optimx package. c("subplex", "pracmanm", "anms") by default.}
}
\value{
data.frame
}
\description{
Canopy Modelling
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(22, 40),
  correct_max = TRUE,
  add_zero = TRUE
)
names(out)
plot(out, c(22, 40))
out$param
}
