% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_predict.R
\name{coef.modeler_HTP}
\alias{coef.modeler_HTP}
\title{Coefficients of an object of class \code{modeler_HTP}}
\usage{
\method{coef}{modeler_HTP}(x, plot_id = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler_HTP} resulting of
executing the function \code{modeler_HTP()}}

\item{plot_id}{A numeric to filter by Plot Id.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A data.frame object with coefficients and standard errors.
}
\description{
Coefficients for an object of class \code{modeler_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "GLI_2"),
  row = "Row",
  range = "Range"
)
mod <- modeler_HTP(
  x = results,
  index = "Canopy",
  plot_id = 1:2,
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise"
)
mod
coef(mod)
}
\author{
Johan Aparicio [aut]
}
