% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_growth_curves.R
\name{fn_quad}
\alias{fn_quad}
\title{Quadratic function}
\usage{
fn_quad(t, a, b, c)
}
\arguments{
\item{t}{Numeric value.}

\item{a}{Numeric value for coefficient a.}

\item{b}{Numeric value for coefficient b.}

\item{c}{Numeric value for coefficient c.}
}
\value{
A numeric value based on the linear function.
}
\description{
Computes a value based on a quadratic function..
}
\details{
\if{html}{
\deqn{
f(t; a, b, c) = a \cdot t^2 + b \cdot t + c
}
}
}
\examples{
library(flexFitR)
plot_fn(fn = "fn_quad", params = c(a = 1, b = 10, c = 5))
}
