% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plant_height.R
\name{height_HTP}
\alias{height_HTP}
\title{Plant Height Modelling}
\usage{
height_HTP(
  results,
  canopy,
  plant_height = "PH",
  plot_id = NULL,
  add_zero = TRUE,
  method = c("lbfgsb3c", "pracmanm", "anms"),
  return_method = FALSE,
  parameters = c(t2 = 67, alpha = 1/600, beta = -1/80),
  fn = fn_sse_exp
)
}
\arguments{
\item{results}{Object of class exploreHTP}

\item{canopy}{Object of class canopy_HTP}

\item{plant_height}{string}

\item{plot_id}{Optional Plot ID. NULL by default}

\item{add_zero}{TRUE or FALSE. Add zero to the time series.TRUE by default.}

\item{method}{A vector of the methods to be used, each as a character string.
See optimx package. c("subplex", "pracmanm", "anms") by default.}

\item{return_method}{TRUE or FALSE. To return the method selected for the
optimization in the table. FALSE by default.}

\item{parameters}{(Optional)    Initial values for the parameters to be
optimized over. c(45, 80) by default.}

\item{fn}{A function to be minimized (or maximized), with first argument the
vector of parameters over which minimization is to take place.
It should return a scalar result. Default is \link{fn_sse_exp}.}
}
\value{
data.frame
}
\description{
Plant Height Modelling
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(22, 40),
  correct_max = TRUE,
  add_zero = TRUE
)
names(out)
plot(out, c(22, 40))
out$param
}
