% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plant_height.R
\name{height_HTP}
\alias{height_HTP}
\title{Plant Height Modelling}
\usage{
height_HTP(
  results,
  canopy,
  plant_height = "PH",
  plot_id = NULL,
  add_zero = TRUE,
  method = c("subplex", "pracmanm", "anms"),
  return_method = FALSE,
  parameters = c(t2 = 67, alpha = 1/600, beta = -1/80),
  fn = sse_exp2_exp
)
}
\arguments{
\item{results}{Object of class exploreHTP}

\item{canopy}{Object of class canopy_HTP}

\item{plant_height}{A string specifying the Plant Height trait to be modeled. Default is \code{"PH"}.}

\item{plot_id}{Optional Plot ID. NULL by default}

\item{add_zero}{TRUE or FALSE. Add zero to the time series.TRUE by default.}

\item{method}{A vector of the methods to be used, each as a character string.
See optimx package. c("subplex", "pracmanm", "anms") by default.}

\item{return_method}{TRUE or FALSE. To return the method selected for the
optimization in the table. FALSE by default.}

\item{parameters}{(Optional)    Initial values for the parameters to be
optimized over. c(45, 80) by default.}

\item{fn}{A function to be minimized (or maximized), with first argument the
vector of parameters over which minimization is to take place.
It should return a scalar result. Default is \link{sse_exp2_exp}.}
}
\value{
data.frame
}
\description{
Plant Height Modelling
}
\examples{
library(exploreHTP)
data(dt_chips)
results <- read_HTP(
  data = dt_chips,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(60, 150),
  correct_max = TRUE,
  add_zero = TRUE
)
names(out)
plot(out, plot_id = c(60, 150))
ph_1 <- height_HTP(
  results = results,
  canopy = out,
  plant_height = "PH",
  add_zero = TRUE,
  method = c("nlminb", "anms", "mla", "pracmanm", "subplex"),
  return_method = TRUE,
  parameters = c(t2 = 67, alpha = 1 / 600, beta = -1 / 80),
  fn = sse_exp2_exp
)
plot(
  x = ph_1,
  plot_id = c(60, 150),
  fn = quote(fn_exp2_exp(time, t1, t2, alpha, beta))
)
ph_1$param

ph_2 <- height_HTP(
  results = results,
  canopy = out,
  plant_height = "PH",
  add_zero = TRUE,
  method = c("nlminb", "anms", "mla", "pracmanm", "subplex"),
  return_method = TRUE,
  parameters = c(t2 = 67, alpha = 1 / 600, beta = -1 / 80),
  fn = sse_exp2_lin
)
plot(
  x = ph_2,
  plot_id = c(60, 150),
  fn = quote(fn_exp2_lin(time, t1, t2, alpha, beta))
)
ph_2$param
}
