% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_growth_curves.R
\name{list_methods}
\alias{list_methods}
\title{Print available methods in flexFitR}
\usage{
list_methods(bounds = FALSE, check_package = FALSE)
}
\arguments{
\item{bounds}{If TRUE, returns methods for box (or bounds) constraints. FALSE  by default.}

\item{check_package}{If TRUE, ensures solvers are installed. FALSE  by default.}
}
\value{
A vector with available methods
}
\description{
Print available methods in flexFitR
}
\examples{
library(flexFitR)
list_methods()
}
