% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_maturity.R
\name{maturity_HTP}
\alias{maturity_HTP}
\title{Maturity Modelling}
\usage{
maturity_HTP(
  results,
  canopy,
  index = "GLI",
  check_negative = TRUE,
  add_zero = TRUE,
  plot_id = NULL,
  method = c("subplex", "pracmanm", "anms"),
  return_method = FALSE,
  parameters = c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01),
  fn = sse_lin_pl_lin
)
}
\arguments{
\item{results}{Object of class exploreHTP}

\item{canopy}{Object of class canopy_HTP}

\item{index}{A string specifying the trait to be modeled. Default is \code{"GLI"}.}

\item{check_negative}{TRUE of FALSE. Convert negative values to zero.}

\item{add_zero}{TRUE or FALSE. Add zero to the time series.TRUE by default.}

\item{plot_id}{Optional Plot ID. NULL by default}

\item{method}{A vector of the methods to be used, each as a character string.
See optimx package. c("subplex", "pracmanm", "anms") by default.}

\item{return_method}{TRUE or FALSE. To return the method selected for the
optimization in the table. FALSE by default.}

\item{parameters}{(Optional)    Initial values for the parameters to be
optimized over. c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01) by default.}

\item{fn}{A function to be minimized (or maximized), with first argument the
vector of parameters over which minimization is to take place.
It should return a scalar result. Default is \link{sse_lin_pl_lin}.}
}
\value{
data.frame
}
\description{
Maturity Modelling
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "GLI_2"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(195, 40),
  correct_max = TRUE,
  add_zero = TRUE
)
mat <- maturity_HTP(
  results = results,
  canopy = out,
  index = "GLI_2",
  parameters = c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01),
  fn = sse_lin_pl_lin
)
plot(mat, plot_id = c(195, 40))
}
