% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_maturity.R
\name{maturity_HTP}
\alias{maturity_HTP}
\title{Maturity Modelling}
\usage{
maturity_HTP(x, index = "GLI", canopy = "Canopy", plot_id = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{read_HTP}, containing the results of the \code{read_HTP()} function.}

\item{index}{A string specifying the canopy trait to be modeled. Default is \code{"GLI"}.}

\item{canopy}{A string specifying the canopy trait to be modeled. Default is \code{"Canopy"}.}

\item{plot_id}{An optional vector of plot IDs to filter the data. Default is \code{NULL}, meaning all plots are used.}

\item{...}{Additional arguments passed to the \code{modeler_HTP()} function.}
}
\value{
An object of class \code{modeler_HTP}, which is a list containing the following elements:
\describe{
  \item{\code{param}}{A data frame containing the optimized parameters and related information.}
  \item{\code{dt}}{A data frame with the corrected and possibly zero-augmented canopy data.}
  \item{\code{fn}}{The call used to calculate the AUC.}
  \item{\code{max_time}}{Maximum time value used for calculating the AUC.}
}
}
\description{
Maturity Modelling
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "GLI_2"),
  row = "Row",
  range = "Range"
)
mat <- maturity_HTP(
  x = results,
  index = "GLI_2",
  canopy = "Canopy",
  plot_id = c(195, 40)
)
plot(mat, plot_id = c(195, 40))
print(mat)
}
