% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.canopy_HTP}
\alias{plot.canopy_HTP}
\title{Plot an object of class \code{canopy_HTP}}
\usage{
\method{plot}{canopy_HTP}(x, plot_id = NULL, label_size = 4, base_size = 14, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{canopy_HTP} resulting of
executing the function \code{canopy_HTP()}}

\item{plot_id}{To avoid too many plots in one figure. Filter by Plot Id.}

\item{label_size}{Numeric. Size of the labels in the plot. Default is 4.}

\item{base_size}{Base font size, given in pts. Default is 14.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{canopy_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
names(results)
out <- canopy_HTP(
  results = results,
  canopy = "Canopy",
  plot_id = c(22, 40),
  correct_max = TRUE,
  add_zero = TRUE
)
names(out)
plot(out, c(22, 40))
out$param$deltaT <- out$param$t2 - out$param$t1
out$param$slope <- out$param$k / out$param$deltaT
out$param
}
\author{
Johan Aparicio [aut]
}
