% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.height_HTP}
\alias{plot.height_HTP}
\title{Plot an object of class \code{height_HTP}}
\usage{
\method{plot}{height_HTP}(
  x,
  plot_id = NULL,
  label_size = 4,
  base_size = 14,
  fn = quote(fn_exp_linear(time, t1, t2, alpha, beta)),
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{height_HTP} resulting of
executing the function \code{height_HTP()}}

\item{plot_id}{To avoid too many plots in one figure. Filter by Plot Id.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{fn}{Object of class call. e.g. \code{quote(fn_exp_linear(time, t1, t2, alpha, beta))}}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{height_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
out <- canopy_HTP(results, plot_id = 22)
plot(out)
out$param
}
\author{
Johan Aparicio [aut]
}
