% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.maturity_HTP}
\alias{plot.maturity_HTP}
\title{Plot an object of class \code{maturity_HTP}}
\usage{
\method{plot}{maturity_HTP}(
  x,
  plot_id = NULL,
  label_size = 4,
  base_size = 14,
  fn = quote(fn_lin_pl_lin(time, t1, t2, t3, k, beta)),
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{maturity_HTP} resulting of
executing the function \code{maturity_HTP()}}

\item{plot_id}{To avoid too many plots in one figure. Filter by Plot Id.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{fn}{Object of class call. e.g. \code{quote(fn_lin_pl_lin(time, t1, t2, t3, k, beta))}}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{maturity_HTP}
}
\examples{
# in progress
}
\author{
Johan Aparicio [aut]
}
