% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_plots.R
\name{plot.modeler}
\alias{plot.modeler}
\title{Plot an object of class \code{modeler}}
\usage{
\method{plot}{modeler}(x, id = NULL, type = 1, label_size = 4, base_size = 14, parm = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{id}{To avoid too many plots in one figure. Filter by group Id.}

\item{type}{Numeric 1, 2, or 3. To specify the type of plot. Default is 1.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{parm}{If type is equal to 2 it must be a vector of names of the parameters. If NULL, all parameters are considered.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{modeler}
}
\examples{
library(exploreHTP)
data(dt_potato)
explorer <- explorer(dt_potato, x = DAP, y = c(Canopy, GLI_2), id = Plot)
# Example 1
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    id = c(1:3),
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    add_zero = TRUE,
    max_as_last = TRUE
  )
mod_1
plot(mod_1, id = 1:2)
plot(mod_1, id = 1:3, type = 2, label_size = 10)
}
\author{
Johan Aparicio [aut]
}
