% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_plots.R
\name{plot.modeler}
\alias{plot.modeler}
\title{Plot an object of class \code{modeler}}
\usage{
\method{plot}{modeler}(
  x,
  id = NULL,
  type = 1,
  label_size = 4,
  base_size = 14,
  color = "red",
  parm = NULL,
  n_points = 2000,
  title = NULL,
  add_ci = TRUE,
  add_ribbon = FALSE,
  color_ribbon = "blue",
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{id}{To avoid too many plots in one figure. Filter by group Id.}

\item{type}{Numeric 1, 2, or 3. To specify the type of plot. Default is 1.
 Type 4, 5 and 6 are experimental.
\describe{
  \item{\code{type = 1}}{Raw Data + Fitted Curve}
  \item{\code{type = 2}}{Coefficients + Confindence Intervals}
  \item{\code{type = 3}}{Fitted Curve + Color by Group}
  \item{\code{type = 4}}{Fitted curve + Confindence Intervals}
  \item{\code{type = 5}}{First Derivative + Confindence Intervals}
  \item{\code{type = 6}}{Second Derivative + Confindence Intervals}
}}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{color}{color for geom_line when type 1. Default is "red".}

\item{parm}{If type is equal to 2 it must be a vector of names of the parameters. If NULL, all parameters are considered.}

\item{n_points}{Number of points to interpolate along the x axis. Default is 2000.}

\item{title}{Optional string character to add a title to the plot.}

\item{add_ci}{Add confidence interval for type in c(4, 5, 6).TRUE or FALSE. TRUE by default.}

\item{add_ribbon}{Add geom_ribbon for type in c(4, 5, 6).TRUE or FALSE. TRUE by default.}

\item{color_ribbon}{Ribbon color. "blue" by default.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{modeler}
}
\examples{
library(flexFitR)
data(dt_potato)
explorer <- explorer(dt_potato, x = DAP, y = c(Canopy, GLI_2), id = Plot)
# Example 1
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    subset = c(1:3),
    add_zero = TRUE,
    max_as_last = TRUE
  )
print(mod_1)
plot(mod_1, id = 1:2)
plot(mod_1, id = 1:3, type = 2, label_size = 10)
}
\author{
Johan Aparicio [aut]
}
