% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_plots.R
\name{plot.modeler_HTP}
\alias{plot.modeler_HTP}
\title{Plot an object of class \code{modeler_HTP}}
\usage{
\method{plot}{modeler_HTP}(x, id = NULL, label_size = 4, base_size = 14, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler_HTP} resulting of
executing the function \code{modeler_HTP()}}

\item{id}{To avoid too many plots in one figure. Filter by Id.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{modeler_HTP}
}
\examples{
library(exploreHTP)
suppressMessages(library(dplyr))
data(dt_potato)
results <- dt_potato |>
  read_HTP(
    x = DAP,
    y = c(Canopy, GLI_2),
    id = Plot,
    .keep = c(Gen, Row, Range)
  )
names(results)
# Example 1
mod_1 <- modeler_HTP(
  x = results,
  variable = "GLI_2",
  id = c(195),
  parameters = c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01),
  fn = "fn_lin_pl_lin",
)
plot(mod_1, plot_id = c(195))
mod_1
# Example 2
mod_2 <- modeler_HTP(
  x = results,
  variable = "Canopy",
  id = c(195),
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise"
)
plot(mod_2, plot_id = c(195))
mod_2
# Example 3
fixed_params <- results$dt_long |>
  filter(var \%in\% "Canopy") |>
  group_by(uid) |>
  summarise(k = max(y, na.rm = TRUE), .groups = "drop")
mod_3 <- modeler_HTP(
  x = results,
  variable = "Canopy",
  id = c(195),
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise",
  fixed_params = fixed_params
)
plot(mod_3, id = c(195))
mod_3
}
\author{
Johan Aparicio [aut]
}
