% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.modeler_HTP}
\alias{plot.modeler_HTP}
\title{Plot an object of class \code{modeler_HTP}}
\usage{
\method{plot}{modeler_HTP}(x, plot_id = NULL, label_size = 4, base_size = 14, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{modeler_HTP} resulting of
executing the function \code{modeler_HTP()}}

\item{plot_id}{To avoid too many plots in one figure. Filter by Plot Id.}

\item{label_size}{Label size. 3 by default.}

\item{base_size}{Base font size, given in pts.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object.
}
\description{
Create several plots for an object of class \code{modeler_HTP}
}
\examples{
library(exploreHTP)
suppressMessages(library(dplyr))
data(dt_potato)
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "GLI_2"),
  row = "Row",
  range = "Range"
)
names(results)
mat <- modeler_HTP(
  x = results,
  index = "GLI_2",
  plot_id = c(195),
  parameters = c(t1 = 38.7, t2 = 62, t3 = 90, k = 0.32, beta = -0.01),
  fn = "fn_lin_pl_lin",
)
plot(mat, plot_id = c(195))
mat

can <- modeler_HTP(
  x = results,
  index = "Canopy",
  plot_id = c(195),
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise"
)
plot(can, plot_id = c(195))
can

fixed_params <- results$dt_long |>
  filter(trait \%in\% "Canopy") |>
  group_by(plot, genotype) |>
  summarise(k = max(value, na.rm = TRUE), .groups = "drop")
can <- modeler_HTP(
  x = results,
  index = "Canopy",
  plot_id = c(195),
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise",
  fixed_params = fixed_params
)
plot(can, plot_id = c(195))
can
}
\author{
Johan Aparicio [aut]
}
