% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods.R
\name{plot.read_HTP}
\alias{plot.read_HTP}
\title{Plot an object of class \code{read_HTP}}
\usage{
\method{plot}{read_HTP}(
  x,
  type = "trait_by_time",
  label_size = 4,
  signif = FALSE,
  method = "pearson",
  n_row = NULL,
  n_col = NULL,
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{read_HTP} resulting of
executing the function \code{read_HTP()}}

\item{type}{Character string "trait_by_time" or "time_by_trait".}

\item{label_size}{Label size. 4 by default.}

\item{signif}{TRUE or FALSE. Add the pvalue to the correlations plot.}

\item{method}{method="pearson" is the default value. The alternatives to be
passed to cor are "spearman" and "kendall". These last two are much slower,
particularly for big data sets.}

\item{n_row}{Number of rows to use in face_wrap(). NULL by default.}

\item{n_col}{Number of columns to use in face_wrap(). NULL by default.}

\item{...}{Further graphical parameters. For future improvements.}
}
\value{
A ggplot object and an invisible data.frame
}
\description{
Create several plots for an object of class \code{read_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
table <- plot(results, label_size = 4, signif = TRUE, n_row = 2)
table
plot(results, type = "time_by_trait", label_size = 4, signif = TRUE)
}
\author{
Johan Aparicio [aut]
}
