% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_plots.R
\name{plot.read_HTP}
\alias{plot.read_HTP}
\title{Plot an object of class \code{read_HTP}}
\usage{
\method{plot}{read_HTP}(
  x,
  type = "trait_by_time",
  label_size = 4,
  signif = FALSE,
  method = "pearson",
  filter_trait = NULL,
  id = NULL,
  n_row = NULL,
  n_col = NULL,
  base_size = 13,
  return_gg = FALSE,
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{read_HTP}, resulting from executing the function \code{read_HTP()}.}

\item{type}{Character string or number specifying the type of plot to generate. Available options are:
\describe{
  \item{\code{"trait_by_time" or 1}}{Plots correlations between traits over time (default).}
  \item{\code{"time_by_trait" or 2}}{Plots correlations between time points for each trait.}
  \item{\code{"evolution" or 3}}{Plots the evolution of traits over time.}
}}

\item{label_size}{Numeric. Size of the labels in the plot. Default is 4. Only works with type 1 and 2.}

\item{signif}{Logical. If \code{TRUE}, adds p-values to the correlation plot labels. Default is \code{FALSE}. Only works with type 1 and 2.}

\item{method}{Character string specifying the method for correlation calculation. Available options are \code{"pearson"} (default), \code{"spearman"}, and \code{"kendall"}. Only works with type 1 and 2.}

\item{filter_trait}{Character vector specifying the traits to exclude from the plot.}

\item{id}{Optional unique identifier to filter the evolution type of plot. Default is \code{NULL}. Only works with type 3.}

\item{n_row}{Integer specifying the number of rows to use in \code{facet_wrap()}. Default is \code{NULL}. Only works with type 1 and 2.}

\item{n_col}{Integer specifying the number of columns to use in \code{facet_wrap()}. Default is \code{NULL}. Only works with type 1 and 2.}

\item{base_size}{Numeric. Base font size for the plot. Default is 13.}

\item{return_gg}{Logical. If \code{TRUE}, returns the ggplot object instead of printing it. Default is \code{FALSE}.}

\item{...}{Further graphical parameters for future improvements.}
}
\value{
A ggplot object and an invisible data.frame containing the correlation table when \code{type} is \code{"trait_by_time"} or \code{"time_by_trait"}.
}
\description{
Creates various plots for an object of class \code{read_HTP}. Depending on the specified type, the function can generate plots that show correlations between traits over time, correlations between time points for each trait, or the evolution of traits over time.
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(dt_potato, x = DAP, y = c(Canopy, PH), id = Plot)
table <- plot(results, label_size = 4, signif = TRUE, n_row = 2)
table
plot(results, type = "time_by_trait", label_size = 4, signif = TRUE)
}
