% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predict.R
\name{predict.modeler}
\alias{predict.modeler}
\title{Predict an object of class \code{modeler}}
\usage{
\method{predict}{modeler}(object, x = NULL, id = NULL, metadata = FALSE, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{x}{A numeric time point to make the prediction. Can be more than one.}

\item{id}{A numeric to filter by unique identifier. NULL by default.}

\item{metadata}{TRUE or FALSE. Whether to bring the metadata or not when calculating the coefficients.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A data.frame object with predicted values.
}
\description{
Model predictions for an object of class \code{modeler}
}
\examples{
library(exploreHTP)
data(dt_potato)
explorer <- explorer(dt_potato, x = DAP, y = c(Canopy, GLI_2), id = Plot)
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    id = c(15, 2, 45),
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    fn = "fn_piwise",
    add_zero = TRUE,
    max_as_last = TRUE
  )
mod_1
predict(mod_1, x = 45, id = 2)
}
\author{
Johan Aparicio [aut]
}
