% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_predict.R
\name{predict.modeler}
\alias{predict.modeler}
\title{Predict an object of class \code{modeler}}
\usage{
\method{predict}{modeler}(
  object,
  x = NULL,
  id = NULL,
  type = c("point", "auc", "fd", "sd"),
  se_interval = c("confidence", "prediction"),
  n_points = 1000,
  metadata = FALSE,
  ...
)
}
\arguments{
\item{object}{An object inheriting from class \code{modeler} resulting of
executing the function \code{modeler()}}

\item{x}{A numeric time point to make the prediction. Can be more than one.
If type = "auc", x needs to be of size 2, which specifies the interval for
calculating the AUC.}

\item{id}{A unique identifier to filter by. \code{NULL} by default.}

\item{type}{A character string specifying the type of prediction. If "point"
it will predict the value of y for a given x. If "auc" it will return the
area under the curve (AUC) for the fitted curve. If "fd" the first order
derivative at the x value is returned. If "sd" the second order derivative
at the x value is returned.}

\item{se_interval}{Type of standard error calculation for intervals.
"confidence" or "prediction". "confidence" by default.}

\item{n_points}{An integer specifying the number of x points to use for
approximating the Area Under the Curve (AUC). Default is \code{1000}.}

\item{metadata}{\code{TRUE} or \code{FALSE}. Whether to bring the metadata or
not when making the predictions.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A data.frame object with predicted values and standard errors.
}
\description{
Model predictions for an object of class \code{modeler}
}
\examples{
library(flexFitR)
data(dt_potato)
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    subset = c(15, 2, 45),
    add_zero = TRUE,
    max_as_last = TRUE
  )
print(mod_1)
# Point Prediction
predict(mod_1, x = 45, type = "point", id = 2)
# AUC Prediction
predict(mod_1, x = c(0, 108), type = "auc", id = 2)
# First Derivative
predict(mod_1, x = 45, type = "fd", id = 2)
# Second Derivative
predict(mod_1, x = 45, type = "sd", id = 2)
}
\author{
Johan Aparicio [aut]
}
