% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_predict.R
\name{predict.modeler_HTP}
\alias{predict.modeler_HTP}
\title{Predict an object of class \code{modeler_HTP}}
\usage{
\method{predict}{modeler_HTP}(object, x = NULL, id = NULL, metadata = FALSE, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{modeler_HTP} resulting of
executing the function \code{modeler_HTP()}}

\item{x}{A numeric time point to make the prediction. Can be more than one.}

\item{id}{A numeric to filter by unique identifier. NULL by default.}

\item{metadata}{TRUE or FALSE. Whether to bring the metadata or not when calculating the coefficients.}

\item{...}{Further parameters. For future improvements.}
}
\value{
A data.frame object with predicted values.
}
\description{
Model predictions for an object of class \code{modeler_HTP}
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- dt_potato |>
  read_HTP(
    x = DAP,
    y = c(Canopy, GLI_2),
    id = Plot,
    .keep = c(Gen, Row, Range)
  )
mod <- modeler_HTP(
  x = results,
  index = "Canopy",
  id = c(15, 2, 45),
  parameters = c(t1 = 45, t2 = 80, k = 0.9),
  fn = "fn_piwise"
)
mod
predict(mod, x = 45, id = 2)
}
\author{
Johan Aparicio [aut]
}
