% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_S3_methods_print.R
\name{print.modeler_HTP}
\alias{print.modeler_HTP}
\title{Print an object of class \code{modeler_HTP}}
\usage{
\method{print}{modeler_HTP}(x, ...)
}
\arguments{
\item{x}{An object fitted with the function \code{modeler_HTP()}.}

\item{...}{Options used by the tibble package to format the output. See
`tibble::print()` for more details.}
}
\value{
an object inheriting from class \code{modeler_HTP}.
}
\description{
Prints information about \code{modeler_HTP} function.
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
out <- canopy_HTP(x = results, index = "Canopy", plot_id = c(22, 40))
plot(out, plot_id = c(22, 40))
print(out)
}
\author{
Johan Aparicio [aut]
}
