% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read.R
\name{read_HTP}
\alias{read_HTP}
\title{Read HTP data}
\usage{
read_HTP(data, genotype, time, plot, traits, row, range)
}
\arguments{
\item{data}{A data.frame in a wide format.}

\item{genotype}{A character string indicating the column in data that
contains genotypes.}

\item{time}{A character string indicating the column in data that contains the
time points.}

\item{plot}{A character string indicating the column in data that contains the
plot IDs.}

\item{traits}{A character vector specifying the traits for which the models
should be fitted.}

\item{row}{A character string indicating the column in data that contains the
row coordinates.}

\item{range}{A character string indicating the column in data that contains the
range coordinates.}
}
\value{
An object of class \code{read_HTP}, with a list of:
\item{summ_traits}{A data.frame containing a summary of the traits.}
\item{exp_design_resum}{A data.frame containing a summary of the experimental
 design.}
\item{locals_min_max}{A data.frame containing local maximums and minimuns}
\item{dt_long}{A data.frame in long format.}
}
\description{
Read HTP data
}
\examples{
library(exploreHTP)
data(dt_potato)
dt_potato <- dt_potato
results <- read_HTP(
  data = dt_potato,
  genotype = "Gen",
  time = "DAP",
  plot = "Plot",
  traits = c("Canopy", "PH"),
  row = "Row",
  range = "Range"
)
names(results)

head(results$summ_traits)
plot(results, label_size = 4, signif = TRUE, n_row = 2)
}
