% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read.R
\name{read_HTP}
\alias{read_HTP}
\title{Read HTP Data}
\usage{
read_HTP(data, x, y, id = NULL, .keep)
}
\arguments{
\item{data}{A data.frame in a wide format containing HTP data.}

\item{x}{The name of the column in `data` that contains time points.}

\item{y}{The names of the columns in `data` that contain the variables to be analyzed.}

\item{id}{The name of the column in `data` that contains IDs or unique identifiers.}

\item{.keep}{The names of the columns in `data` to keep across the analysis.}
}
\value{
An object of class \code{read_HTP}, which is a list containing the following elements:
\describe{
  \item{\code{summ_traits}}{A data.frame containing summary statistics for each trait at each time point, including minimum, mean, median, maximum, standard deviation, coefficient of variation, number of non-missing values, percentage of missing values, and percentage of negative values.}
  \item{\code{metadata}}{A data.frame summarizing the metadata.}
  \item{\code{locals_min_max}}{A data.frame containing the local minima and maxima of the mean y values over time.}
  \item{\code{dt_long}}{A data.frame in long format, with columns for uid, .keep, var, x, and y}
  \item{\code{.keep}}{A character vector with the names of the variables to keep across.}
}
}
\description{
Reads and processes high-throughput phenotyping (HTP) data from a data frame in wide format.
}
\details{
This function processes and prepares data to be analyzed with \code{modeler_HTP()}.
}
\examples{
library(exploreHTP)
data(dt_potato)
results <- dt_potato |>
  read_HTP(
    x = DAP,
    y = c(Canopy, PH),
    id = Plot,
    .keep = c(Gen, Row, Range)
  )
names(results)
head(results$summ_traits)
plot(results, label_size = 4, signif = TRUE, n_row = 2)
# New data format
head(results$dt_long)
}
